requirements.txt
Alchy
Click<7
coloredlogs
marshmallow
ruamel.yaml
SQLAlchemy
rich